requirements.txt
azure-functions
azure-cosmos>=4.7.0
python-dotenv>=1.0.1
